Part IV/app/Z_Retired handler and serverless/requirements.txt
#requirements warmfuzzy v3
pillow #image processing lib
boto3 #python sdk
pydantic